pyproject.toml
[tool.poetry]
name = "fastapi-rest-contact-book"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
alembic = "1.13.1"
annotated-types = "0.6.0"
anyio = "4.3.0"
async-timeout = "4.0.3"
asyncpg = "0.29.0"
bcrypt = "4.1.2"
cffi = "1.16.0"
click = "8.1.7"
colorama = "0.4.6"
cryptography = "42.0.5"
dnspython = "2.6.1"
ecdsa = "0.18.0"
email-validator = "2.1.1"
greenlet = "3.0.3"
h11 = "0.14.0"
idna = "3.6"
libgravatar = "1.0.4"
mako = "1.3.2"
markupsafe = "2.1.5"
passlib = "1.7.4"
pyasn1 = "0.6.0"
pycparser = "2.22"
pydantic = "2.6.4"
pydantic-core = "2.16.3"
python-jose = "3.3.0"
python-multipart = "0.0.9"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.29"
starlette = "0.37.2"
typing-extensions = "4.10.0"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.1"
redis = "^5.0.3"
cloudinary = "^1.39.1"
fastapi-limiter = "^0.1.6"
pytest = "^8.1.1"
httpx = "^0.27.0"
pytest-mock = "^3.14.0"
scramp = "^1.4.5"

[tool.poetry.group.dev.dependencies]
sphinx = "^7.3.5"

[tool.poetry.group.test.dependencies]
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules -p pytest_asyncio"
testpaths = ["tests"]
pythonpath = "."
